+++ src/negotiation/execution_params.spec.ts
@@ -7,8 +7,8 @@ const defaultExecutionParams = () => {
       peer_id: "QmWsfYSJ1fzeCWG7N3y2RPQMrWNXtqWPWqKqqKbXQq2MMs",
       address_hint: "/ip4/a.b.c.d/tcp/9939"
     },
-    alpha_expiry: moment().unix() + 48 * 60 * 60,
-    beta_expiry: moment().unix() + 24 * 60 * 60,
+    alpha_expiry: moment().unix() + 24 * 60 * 60,
+    beta_expiry: moment().unix() + 12 * 60 * 60,
     ledgers: {
       bitcoin: { network: "mainnet" },
       ethereum: { chain_id: 1 }
@@ -37,11 +37,21 @@ describe("Taker Negotiator", () => {
     expect(validateExecutionParams(executionParams)).toBeTruthy();
   });
 
-  it("Invalidates mainnet execution parameters due to expiry", () => {
+  it("Invalidates mainnet execution parameters due to expiry length", () => {
     const executionParams: ExecutionParams = {
       ...defaultExecutionParams(),
       alpha_expiry: moment().unix() + 2 * 60 * 60,
       beta_expiry: moment().unix() + 1 * 60 * 60
+    };
+
+    expect(validateExecutionParams(executionParams)).toBeFalsy();
+  });
+
+  it("Invalidates mainnet execution parameters due to expiry order", () => {
+    const executionParams: ExecutionParams = {
+      ...defaultExecutionParams(),
+      alpha_expiry: moment().unix() + 48 * 60 * 60,
+      beta_expiry: moment().unix() + 72 * 60 * 60
     };
 
     expect(validateExecutionParams(executionParams)).toBeFalsy();